manufacturing-backend/pyproject.toml
[project]
name = "manufacturing-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.0",
    "boto3>=1.34.0",
    "fastapi>=0.111.0",
    "google-generativeai>=0.7.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pypdf2>=3.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "rq>=1.16.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.29.0",
]
